backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
python-multipart==0.0.5
pillow==8.3.1
markdown==3.3.4
python-dotenv==0.19.0